Navigation as a separate component
